made killing all processes work

+++ source/scripts/os/shell.ts
@@ -168,7 +168,7 @@ module TSOS {
             // kill <id> - kills the specified process id.
             sc = new ShellCommand(this.shellKillProcess,
                                   "kill",
-                                  "<int> - Kill the specified process if it is running.");
+                                  "<int,*> - Kill the specified process if it is running. * kills all.");
             this.commandList[this.commandList.length] = sc;    
 
             // format  -Initialize  all blocks  in  all sectors in  all tracks and display a  message denoting    success or  failure.    
@@ -693,16 +693,20 @@ module TSOS {
 
         }
         public shellKillProcess(args):void{
-            var program = parseInt(args[0]);
-            if (_ExecutingProgramPID !==program && !_Scheduler.readyQueue.inQueue(program)){
+
+            var program = (args[0]!=="*")? parseInt(args[0]):-1;
+            if (_ExecutingProgramPID !==program && !_Scheduler.readyQueue.inQueue(program) &&args[0]!=='*'){
                 if (_SarcasticMode)
                     _StdOut.putText("I cannot kill what which has no life.");
                 else
                     _StdOut.putText("Process cannot be killed because process is not running.");
             }
+            else if (program===-1){
+              _KernelInterruptQueue.enqueue(new Interrupt(PROCESS_KILLED_IRQ, -1));
+            }
             else {
-                _Scheduler.stopRunning(program);
-                _StdOut.putText("PID " + args[0] +" is running.");
+                _KernelInterruptQueue.enqueue(new Interrupt(PROCESS_KILLED_IRQ, program));
+                _StdOut.putText("PID " + args[0] +" is dead.");
             }
         }
         public shellSetScheduling(args):void{
